Opdracht 4

+++ Opdracht 2.py
@@ -7,13 +7,13 @@ AantalVuurStenen = 4
 AantalSisatouw = 1
 AantalPersonen = 4
 
-PrijsPemmicanRepen = int(input("Hoeveel kost een pakje Pemmican-Repen? "))
-PrijsVijgenPlak = int(input("Hoeveel kost een VijgenPlak? "))
-PrijsScheepsBiscuitRol = int(input("Hoeveel kost een scheepsbiscuit rol? "))
-PrijsFireSteel = int(input("Hoeveel kost een firesteel? "))
-PrijsPakjeLucifers = int(input("Hoeveel kost een pakje lucifers? "))
-PrijsVuurSteen = int(input("Hoeveel kost een originele vuursteen? "))
-PrijsSisalTouw2Meter = int(input("Hoeveel kost een sisaltouw per 2 meter? "))
+PrijsPemmicanRepen = float(input("Hoeveel kost een pakje Pemmican-Repen? "))
+PrijsVijgenPlak = float(input("Hoeveel kost een VijgenPlak? "))
+PrijsScheepsBiscuitRol = float(input("Hoeveel kost een scheepsbiscuit rol? "))
+PrijsFireSteel = float(input("Hoeveel kost een firesteel? "))
+PrijsPakjeLucifers = float(input("Hoeveel kost een pakje lucifers? "))
+PrijsVuurSteen = float(input("Hoeveel kost een originele vuursteen? "))
+PrijsSisalTouw2Meter = float(input("Hoeveel kost een sisaltouw per 2 meter? "))
  
 TotaalPrijsPemmicanRepen = PemmicanRepenPersoon * AantalPersonen / 4 * PrijsPemmicanRepen
 TotaalPrijsVijgenPlak = VijgenPlakkenPersoon * 4 * PrijsVijgenPlak
